Changed uintmax_t back to size_t

+++ MainGame/data/RoomData.hpp
@@ -8,7 +8,7 @@
 #include <vector>
 #include <SFML/System.hpp>
 
-template <uintmax_t size> struct Print;
+template <size_t size> struct Print;
 
 struct GameObjectDescriptor final
 {
