+++ MainGame/utility/grid.hpp
@@ -15,9 +15,9 @@ namespace util
             class iterator final
             {
                 view* ref;
-                uintmax_t i, j;
+                size_t i, j;
 
-                iterator(view *ref, uintmax_t i, uintmax_t j) : ref(ref), i(i), j(j) {}
+                iterator(view *ref, size_t i, size_t j) : ref(ref), i(i), j(j) {}
 
             public:
                 using difference_type = intmax_t;
@@ -67,27 +67,27 @@ namespace util
 
                 iterator operator+(intmax_t val) const
                 {
-                    uintmax_t offset = i + val;
+                    size_t offset = i + val;
                     return iterator(ref, offset % ref->_width, j + offset / ref->_width);
                 }
 
                 iterator& operator+=(intmax_t val)
                 {
-                    uintmax_t offset = i + val;
+                    size_t offset = i + val;
                     i = offset % ref->_width;
                     j += offset / ref->_width;
                     return *this;
                 }
 
                 iterator operator-(intmax_t val) const { return *this + (-val); }
-                iterator& operator-=(uintmax_t val) { return *this += (-val); }
+                iterator& operator-=(size_t val) { return *this += (-val); }
 
                 intmax_t operator-(const iterator& other) const
                 {
                     if (ref != other.ref) return 0;
 
-                    uintmax_t ofs1 = j*ref->_width + i;
-                    uintmax_t ofs2 = other.j*ref->_width + other.i;
+                    size_t ofs1 = j*ref->_width + i;
+                    size_t ofs2 = other.j*ref->_width + other.i;
 
                     return ofs1 - ofs2;
                 }
@@ -112,23 +112,23 @@ namespace util
             using const_iterator = const iterator;
 
             grid* ref;
-            uintmax_t x, y, _width, _height;
+            size_t x, y, _width, _height;
 
-            view (grid* ref, uintmax_t x, uintmax_t y, uintmax_t w, uintmax_t h)
+            view (grid* ref, size_t x, size_t y, size_t w, size_t h)
             : ref(ref), x(x), y(y), _width(w), _height(h) {}
 
         public:
-            T& operator()(uintmax_t i, uintmax_t j) { return (*ref)(x+i, y+j); }
-            const T& operator()(uintmax_t i, uintmax_t j) const { return (*ref)(x+i, y+j); }
+            T& operator()(size_t i, size_t j) { return (*ref)(x+i, y+j); }
+            const T& operator()(size_t i, size_t j) const { return (*ref)(x+i, y+j); }
 
-            T& at(uintmax_t i, uintmax_t j)
+            T& at(size_t i, size_t j)
             {
                 if (i >= _width || j >= _height)
                     throw std::out_of_range("Attempt to access element outside of bounds of the grid!");
                 return operator()(i, j);
             }
 
-            const T& at(uintmax_t i, uintmax_t j) const
+            const T& at(size_t i, size_t j) const
             {
                 if (i >= _width || j >= _height)
                     throw std::out_of_range("Attempt to access element outside of bounds of the grid!");
@@ -158,21 +158,21 @@ namespace util
             friend class grid<T>;
         };
 
-        uintmax_t _width, _height;
+        size_t _width, _height;
         T* elements;
 
     public:
         grid() noexcept : _width(0), _height(0), elements(nullptr) {}
-        grid(uintmax_t w, uintmax_t h) : _width(w), _height(h), elements(new T[w*h]) {}
-        grid(uintmax_t w, uintmax_t h, T* contents) : grid(w, h)
+        grid(size_t w, size_t h) : _width(w), _height(h), elements(new T[w*h]) {}
+        grid(size_t w, size_t h, T* contents) : grid(w, h)
         {
             std::copy(contents, contents+(w*h), elements);
         }
 
         template <typename I>
-        grid(uintmax_t w, uintmax_t h, I begin, I end) : grid(w, h)
+        grid(size_t w, size_t h, I begin, I end) : grid(w, h)
         {
-            uintmax_t i = 0;
+            size_t i = 0;
             for (auto it = begin; it != end; ++it)
             {
                 if (i >= w*h) break;
@@ -201,17 +201,17 @@ namespace util
 
         ~grid() { delete[] elements; }
 
-        T& operator()(uintmax_t i, uintmax_t j) { return elements[j*_width+i]; }
-        const T& operator()(uintmax_t i, uintmax_t j) const { return elements[j*_width+i]; }
+        T& operator()(size_t i, size_t j) { return elements[j*_width+i]; }
+        const T& operator()(size_t i, size_t j) const { return elements[j*_width+i]; }
 
-        T& at(uintmax_t i, uintmax_t j)
+        T& at(size_t i, size_t j)
         {
             if (i >= _width || j >= _height)
                 throw std::out_of_range("Attempt to access element outside of bounds of the grid!");
             return operator()(i, j);
         }
 
-        const T& at(uintmax_t i, uintmax_t j) const
+        const T& at(size_t i, size_t j) const
         {
             if (i >= _width || j >= _height)
                 throw std::out_of_range("Attempt to access element outside of bounds of the grid!");
@@ -230,7 +230,7 @@ namespace util
         T* begin() { return elements; }
         T* end() { return elements+(_width*_height); }
 
-        const view make_view(uintmax_t x, uintmax_t y, uintmax_t _width, uintmax_t _height) const
+        const view make_view(size_t x, size_t y, size_t _width, size_t _height) const
         {
             if (x + _width >= this->_width || y + _height >= this->_height)
                 throw std::out_of_range("Attempt to make a view that spans outside of the bounds of the grid!");
@@ -238,7 +238,7 @@ namespace util
             return view(this, x, y, _width, _height);
         }
 
-        view make_view(uintmax_t x, uintmax_t y, uintmax_t _width, uintmax_t _height)
+        view make_view(size_t x, size_t y, size_t _width, size_t _height)
         {
             if (x + _width >= this->_width || y + _height >= this->_height)
                 throw std::out_of_range("Attempt to make a view that spans outside of the bounds of the grid!");
@@ -246,8 +246,8 @@ namespace util
             return view(this, x, y, _width, _height);
         }
 
-        uintmax_t width() const { return _width; }
-        uintmax_t height() const { return _height; }
+        size_t width() const { return _width; }
+        size_t height() const { return _height; }
     };
 
     template <typename T>
